Remove CHECK_FOR_NAN option.
I am removing this check because I have some many ways to check the
origin of a NaN using the upside_engine library.  It might be useful to
have a special abort on NaN check in the integrator to stop integration as
soon as a NaN occurs in the derivative.  There is no reason to hide this
behind a flag, however, as the user will almost never want to continue
simulation after a NaN is produced.  The only exception I can think of
is if I later add a hybrid Monte Carlo mode, where NaN could merely
trigger rejection.